+++ src/main/kotlin/com/github/gllxl/imagepreview/CssLineMarkerContributor.kt
@@ -15,10 +15,15 @@ class CssLineMarkerContributor : LineMakerContributor() {
 
     val document = PsiDocumentManager.getInstance(element.project).getDocument(element.containingFile) ?: return null
     val virtualFile = element.containingFile.virtualFile ?: return null
-    val lineNumber = document.getLineNumber(element.textOffset)
     val imageService = imagePreviewService(element.project)
     val imageReference = imageService.resolveImageReference(rawImageReference, virtualFile) ?: return null
-    imageService.references.setLineMapping(virtualFile, lineNumber, imageReference)
+    imageService.references.setLineMapping(
+      virtualFile,
+      document,
+      cssUriElement.textRange,
+      rawImageReference,
+      imageReference,
+    )
 
     return getLineMaker(imageReference, element.project)
   }